Cambio id a long y pongo bigdecimal en donacion
Uso long por recomendacion hibernate y cambio a bigdecimal donacion

+++ src/main/java/com/practicasalma/proyectoalma/dao/AlumnoDAO.java
@@ -76,7 +76,7 @@ public class AlumnoDAO {
         }
     }
 
-    public Alumno obtenerCompleto(Integer id) {
+    public Alumno obtenerCompleto(Long id) {
         try (EntityManager em = GestorBBDD.getEntityManagerFactory().createEntityManager()) {
             // Buscamos al alumno por su ID
             Alumno alumno = em.find(Alumno.class, id);
